lab 3 pdav - byte to int problem conversion

+++ Semester5/Audio-video Data Processing/Laboratoy/src/domain/MatrixOperation.java
@@ -1,27 +1,27 @@
 package domain;
 
 class MatrixOperation {
-    static double[][] divideMatrixes(double[][] A, double[][] B) {
-        double[][] result = new double[8][8];
+    static int[][] divideMatrixes(int[][] A, double[][] B) {
+        int[][] result = new int[8][8];
         double aux;
         for (int i = 0; i < 8; i++)
             for (int j = 0; j < 8; j++) {
                 aux = A[i][j] / B[i][j];
                 if (aux < 0)
-                    result[i][j] = Math.ceil(aux);
+                    result[i][j] = (int) Math.ceil(aux);
                 else
-                    result[i][j] = Math.floor(aux);
+                    result[i][j] = (int) Math.floor(aux);
             }
 
         return result;
     }
 
-    static double[][] multiplyMatrixes(double[][] A, double[][] B) {
-        double[][] result = new double[8][8];
+    static int[][] multiplyMatrixes(int[][] A, double[][] B) {
+        int[][] result = new int[8][8];
 
         for (int i = 0; i < 8; i++)
             for (int j = 0; j < 8; j++)
-                result[i][j] = A[i][j] * B[i][j];
+                result[i][j] = (int) (A[i][j] * B[i][j]);
 
         return result;
     }
